+++ packages/atproto/lib/src/services/identity_service.dart
@@ -8,6 +8,7 @@ import 'package:atproto_core/atproto_core.dart' as core;
 // 🌎 Project imports:
 import '../nsids.g.dart' as ns;
 import 'entities/did.dart';
+import 'entities/plc_operation.dart';
 
 /// Represents `com.atproto.identity.*` service.
 final class IdentityService {
@@ -31,6 +32,7 @@ final class IdentityService {
         to: DID.fromJson,
       );
 
+  /// https://atprotodart.com/docs/lexicons/com/atproto/identity/updateHandle
   Future<core.XRPCResponse<core.EmptyData>> updateHandle({
     required String handle,
   }) async =>
@@ -39,6 +41,37 @@ final class IdentityService {
         body: {
           'handle': handle,
         },
+      );
+
+  /// https://atprotodart.com/docs/lexicons/com/atproto/identity/submitPlcOperation
+  Future<core.XRPCResponse<core.EmptyData>> submitPlcOperation(
+    final Map<String, dynamic> operation,
+  ) async =>
+      await _ctx.post(
+        ns.comAtprotoIdentitySubmitPlcOperation,
+        body: {
+          'operation': operation,
+        },
+      );
+
+  /// https://atprotodart.com/docs/lexicons/com/atproto/identity/signPlcOperation
+  Future<core.XRPCResponse<PlcOperation>> signPlcOperation({
+    String? token,
+    List<String>? rotationKeys,
+    String? alsoKnownAs,
+    Map<String, dynamic>? verificationMethods,
+    Map<String, dynamic>? services,
+  }) async =>
+      await _ctx.post(
+        ns.comAtprotoIdentitySignPlcOperation,
+        body: {
+          'token': token,
+          'rotationKeys': rotationKeys,
+          'alsoKnownAs': alsoKnownAs,
+          'verificationMethods': verificationMethods,
+          'services': services,
+        },
+        to: PlcOperation.fromJson,
       );
 
   Future<core.XRPCResponse<T>> _findDID<T>({